Reduce visibility of lowering
